pi-frame/requirements.txt
pygame>=2.5.0
requests>=2.31.0
Pillow>=10.0.0
click>=8.1.0
python-dotenv>=1.0.0
